Adding Authentication to journal endpoints.

+++ src/main/java/com/example/demo/service/JournalEntryService.java
@@ -30,8 +30,7 @@ public class JournalEntryService {
             journalEntry.setDate(LocalDateTime.now());
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
-//            user.setUserName(null);
-            userService.SaveEntry(user);
+            userService.saveUser(user);
         }catch(Exception e ){
             log.error("Exception",e);
 //            System.out.println(e);
@@ -49,9 +48,21 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
+    @Transactional
     public boolean deleteById(ObjectId id, String userName){
-        journalEntryRepository.deleteById(id);
-        return true;
+        boolean removed = false;
+        try{
+            User user = userService.findByuserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if(removed){
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        }catch(Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An error occured while deleting the entry. ",e);
+        }
+        return removed;
     }
